+++ src/core/Configuration.cpp
@@ -467,7 +467,7 @@ UnicodeString __fastcall TConfiguration::BannerHash(const UnicodeString & Banner
   RawByteString Result;
   Result.SetLength(16);
   md5checksum(
-    reinterpret_cast<const char*>(Banner.c_str()), Banner.Length() * sizeof(wchar_t),
+    reinterpret_cast<const char*>(Banner.c_str()), (int)(Banner.Length() * sizeof(wchar_t)),
     (unsigned char*)Result.c_str());
   return BytesToHex(Result);
 }
